+++ src/glum/_glm.py
@@ -83,15 +83,18 @@ ShapedArrayLike = Union[
 ]
 
 
-def check_array_tabmat_compliant(mat: ArrayLike, **kwargs):
+def check_array_tabmat_compliant(mat: ArrayLike, drop_first: int = False, **kwargs):
     to_copy = kwargs.get("copy", False)
 
     if isinstance(mat, pd.DataFrame) and any(mat.dtypes == "category"):
-        mat = tm.from_pandas(mat)
+        mat = tm.from_pandas(mat, drop_first=drop_first)
 
     if isinstance(mat, tm.SplitMatrix):
         kwargs.update({"ensure_min_features": 0})
-        new_matrices = [check_array_tabmat_compliant(m, **kwargs) for m in mat.matrices]
+        new_matrices = [
+            check_array_tabmat_compliant(m, drop_first=drop_first, **kwargs)
+            for m in mat.matrices
+        ]
         new_indices = [elt.copy() for elt in mat.indices] if to_copy else mat.indices
         return tm.SplitMatrix(new_matrices, new_indices)
 
@@ -102,7 +105,7 @@ def check_array_tabmat_compliant(mat: ArrayLike, **kwargs):
 
     if isinstance(mat, tm.StandardizedMatrix):
         return tm.StandardizedMatrix(
-            check_array_tabmat_compliant(mat.mat, **kwargs),
+            check_array_tabmat_compliant(mat.mat, drop_first=drop_first, **kwargs),
             check_array(mat.shift, **kwargs),
         )
 
@@ -207,6 +210,20 @@ def _check_offset(
         raise ValueError("Offsets must have the same length as y.")
 
     return offset
+
+
+def _name_categorical_variables(
+    categories: Tuple[str], column_name: str, drop_first: bool
+):
+    new_names = [
+        f"{column_name}__{category}" for category in categories[int(drop_first) :]
+    ]
+    if len(new_names) == 0:
+        raise ValueError(
+            f"Categorical column: {column_name}, contains only one category. "
+            + "This should be dropped from the feature matrix."
+        )
+    return new_names
 
 
 def check_bounds(
@@ -687,6 +704,7 @@ class GeneralizedLinearRegressorBase(BaseEstimator, RegressorMixin):
         A_ineq: Optional[np.ndarray] = None,
         b_ineq: Optional[np.ndarray] = None,
         force_all_finite: bool = True,
+        drop_first: bool = False,
     ):
         self.l1_ratio = l1_ratio
         self.P1 = P1
@@ -716,6 +734,7 @@ class GeneralizedLinearRegressorBase(BaseEstimator, RegressorMixin):
         self.A_ineq = A_ineq
         self.b_ineq = b_ineq
         self.force_all_finite = force_all_finite
+        self.drop_first = drop_first
 
     @property
     def family_instance(self) -> ExponentialDispersionModel:
@@ -1207,6 +1226,7 @@ class GeneralizedLinearRegressorBase(BaseEstimator, RegressorMixin):
             copy=True,
             ensure_2d=True,
             allow_nd=False,
+            drop_first=self.drop_first,
         )
 
         if alpha_index is None:
@@ -1279,6 +1299,7 @@ class GeneralizedLinearRegressorBase(BaseEstimator, RegressorMixin):
             copy=self._should_copy_X(),
             ensure_2d=True,
             allow_nd=False,
+            drop_first=self.drop_first,
         )
         eta = self.linear_predictor(
             X, offset=offset, alpha_index=alpha_index, alpha=alpha
@@ -1720,14 +1741,16 @@ class GeneralizedLinearRegressorBase(BaseEstimator, RegressorMixin):
             if any(X.dtypes == "category"):
                 self.feature_names_ = list(
                     chain.from_iterable(
-                        [f"{column}__{category}" for category in dtype.categories]
+                        _name_categorical_variables(
+                            dtype.categories, column, self.drop_first
+                        )
                         if pd.api.types.is_categorical_dtype(dtype)
                         else [column]
                         for column, dtype in zip(X.columns, X.dtypes)
                     )
                 )
 
-                def _expand_categorical_penalties(penalty, X):
+                def _expand_categorical_penalties(penalty, X, drop_first):
                     """
                     If P1 or P2 has the same shape as X before expanding the
                     categoricals, we assume that the penalty at the location of
@@ -1749,7 +1772,7 @@ class GeneralizedLinearRegressorBase(BaseEstimator, RegressorMixin):
                         return np.array(
                             list(
                                 chain.from_iterable(
-                                    [elmt for _ in dtype.categories]
+                                    [elmt for _ in dtype.categories[int(drop_first) :]]
                                     if pd.api.types.is_categorical_dtype(dtype)
                                     else [elmt]
                                     for elmt, dtype in zip(penalty, X.dtypes)
@@ -1759,10 +1782,10 @@ class GeneralizedLinearRegressorBase(BaseEstimator, RegressorMixin):
                     else:
                         return penalty
 
-                P1 = _expand_categorical_penalties(self.P1, X)
-                P2 = _expand_categorical_penalties(self.P2, X)
+                P1 = _expand_categorical_penalties(self.P1, X, self.drop_first)
+                P2 = _expand_categorical_penalties(self.P2, X, self.drop_first)
 
-                X = tm.from_pandas(X)
+                X = tm.from_pandas(X, drop_first=self.drop_first)
             else:
                 self.feature_names_ = X.columns
 
@@ -1799,6 +1822,7 @@ class GeneralizedLinearRegressorBase(BaseEstimator, RegressorMixin):
                 dtype=_dtype,
                 copy=copy_X,
                 force_all_finite=force_all_finite,
+                drop_first=self.drop_first,
             )
             self._check_n_features(X, reset=True)
         else:
@@ -2109,6 +2133,11 @@ class GeneralizedLinearRegressor(GeneralizedLinearRegressorBase):
         ``A_ineq w <= b_ineq``. Refer to the documentation of ``A_ineq`` for
         details.
 
+    drop_first : bool, optional (default = False)
+        If ``True``, drop the first column when encoding categorical variables.
+        Set this to True when alpha=0 and solver='auto' to prevent an error due to a singular
+        feature matrix.
+
     Attributes
     ----------
     coef_ : numpy.array, shape (n_features,)
@@ -2189,6 +2218,7 @@ class GeneralizedLinearRegressor(GeneralizedLinearRegressorBase):
         A_ineq: Optional[np.ndarray] = None,
         b_ineq: Optional[np.ndarray] = None,
         force_all_finite: bool = True,
+        drop_first: bool = False,
     ):
         self.alphas = alphas
         self.alpha = alpha
@@ -2221,6 +2251,7 @@ class GeneralizedLinearRegressor(GeneralizedLinearRegressorBase):
             A_ineq=A_ineq,
             b_ineq=b_ineq,
             force_all_finite=force_all_finite,
+            drop_first=drop_first,
         )
 
     def _validate_hyperparameters(self) -> None: