+++ src/glum/_glm_cv.py
@@ -242,6 +242,8 @@ class GeneralizedLinearRegressorCV(GeneralizedLinearRegressorBase):
         number of CPU on your machine. ``None`` means ``1`` unless in a
         :obj:`joblib.parallel_backend` context.
 
+    drop_first : bool, optional (default = False)
+        If ``True``, drop the first column when encoding categorical variables.
 
     Attributes
     ----------
@@ -306,6 +308,7 @@ class GeneralizedLinearRegressorCV(GeneralizedLinearRegressorBase):
         force_all_finite: bool = True,
         cv=None,
         n_jobs: Optional[int] = None,
+        drop_first: bool = False,
     ):
         self.alphas = alphas
         self.cv = cv
@@ -338,6 +341,7 @@ class GeneralizedLinearRegressorCV(GeneralizedLinearRegressorBase):
             A_ineq=A_ineq,
             b_ineq=b_ineq,
             force_all_finite=force_all_finite,
+            drop_first=drop_first,
         )
 
     def _validate_hyperparameters(self) -> None: